+++ src/lib/store.ts
@@ -12,7 +12,7 @@ export let state = {
   linkHost: '',
   dlFormat: 'opus',
   theme: 'auto',
-  customTheme: '',
+  customColor: '',
   roundness: '0.4rem',
   searchSuggestions: true,
   searchFilter: '',
@@ -90,7 +90,6 @@ export const store: {
   },
   linkHost: string,
   searchQuery: string,
-  superCollectionType: 'featured' | 'collections' | 'channels' | 'feed' | 'playlists',
   addToCollectionOptions: string[],
   actionsMenu: CollectionItem,
   list: List & Record<'url' | 'type' | 'uploader', string>,
@@ -137,7 +136,6 @@ export const store: {
   },
   linkHost: state.linkHost || location.origin,
   searchQuery: '',
-  superCollectionType: state.defaultSuperCollection as 'featured',
   actionsMenu: {
     id: '',
     title: '',
